Integrate ml model training in the model creation
The ml SavedModel will reside in the RTS model dir and so it should be
generated by the same program that generates the rest of the RTS model
dir. ML training data will still be a separate command similar to
fetching the rejects and durations.

Bug: 1289329
Change-Id: I402e83cd6817b08b61f53fea1b36c8db98ae7b93
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3688842
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Gregory Guterman <guterman@google.com>
Cr-Commit-Position: refs/heads/main@{#49757}