Add UFS Client to Satlab CLI
We will use UFS in the future to help launch servod containers. This CL
just adds a lightweight wrapper for the raw UFS client. Done so a) we
have an interface easier to mock/swap in future and b) to make it more
clear what we do with UFS

Bug=b:232335621

Change-Id: I4a2989ffd730503e70740dff37f092b5356d3ba7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3705596
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Prasad Vuppalapu <prasadv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49791}