[recovery]: Include Stderr from runner execution.
This commit is a part of the PARIS project. It introduces the
inclusion of stderr from the execution of a command through a runner
instance for a DUT.

When a command is executed on a DUT using an instance of a runner, it
can error out. Currently we capture any such error conditions, and
the correspondnig exit code. Additionally, we also need to capture the
std-error. This will allow the caller for the runner to examine the
contents of std-error and use it for things such as setting the state
of DUT etc.

BUG=b:235528471
TEST=build

Change-Id: I6fefe87eb442e42916c0e6b34f8155e82837dcb5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3708048
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#49794}