Update bootstrapper to only ensure CIPD packages once.
The bootstrapper code currently does separate ensure operations for
getting the bootstrapped executable and getting gclient when necessary.
This is incorrect because the ensure operation removes existing contents
to match it to what is requested. This results in a race condition where
the bootstrapped executable is sometimes removed when downloading the
depot tools package.

This change modifies the bootstrapper code to determine up front if the
bootstrap configuration of the build might require the depot tools
package and does a single ensure operation. The downloading of packages
is handled by the DownloadPackages function, which also takes care of
downloading the executable from CAS instead of CIPD if led was used to
override the recipe.

Bug: 1313607
Change-Id: I867b0f30948897141f266431ad0bf810c21321c4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3705925
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#49817}