[UFS] Refactor public service account validation.
Refactoring the service account validation for Public tests as it is not always possible to derive the service account from the auth context. For example CTP binary will need to send the service account as part of the request parameters and this change will make it possible to validate with service accounts obtained from both api request and auth context.

Bug: b/209476191
Test: go test; make dev;
Change-Id: Ie2c7b8bb2cc02fc97c17db4f05e7cab8d424ace6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3717539
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Gowri Denduluri <gowriden@google.com>
Cr-Commit-Position: refs/heads/main@{#49845}