[shivas] Add jsonutil for parsing arrays of JSON
So, parsing a JSON object that might be a proto.Message OR might be an array of
proto.Messages is somewhat tricky.

I think the best path forward is to write a library with a function that can
take a proto.Message as a schema, and then parse a json payload that may be an object
or may be an array of objects, since this will be used in several places.

This will let us make `shivas get [RESOURCE]` and `shivas update [RESOURCE]` round trip
in the expected way.

Tests:

The specific error messages in [1] and [2] indicate that we're successfully parsing both a hostname and a list
of hostnames.

[1] ./dev-shivas add host -namespace browser -f /tmp/a WHERE /tmp/a contains the JSON of a previously deleted host.
{
        "name": "web",
...
REDACTED
...
        "ownership": null
}

[2] $ ./dev-shivas add host -namespace browser -f <(echo '{"name": "hi", "machines": ["hi"], "hostname": "hi"}')
add: rpc error: code = NotFound desc = unable to get machine hi: rpc error: code = NotFound desc ...

[3] $ ./dev-shivas add host -namespace browser -f <(echo '[{"name": "hi", "machines": ["hi"], "hostname": "hi"}'])
add: rpc error: code = NotFound desc = unable to get machine hi: rpc error: code = NotFound desc ...

BUG=b:193566796
TEST=make test AND manual tests described above.

Change-Id: Ia2a6dffb68ed4e028452528e3035e5eb152cde50
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3699084
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Gowri Denduluri <gowriden@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#49879}