[recovery]: Cleaned-up the canary-test to address the issue of failure with setting power state to rec.
Setting the power state to rec can sometimes return error. In legacy
repair, any such errors are not handled. However, in Paris, such
errors cause the recovery task to fail. The previous commit on this
file includes a change to suppress such task failures. Instead, such
errors are logged, and task continues without error.

This commit is a part of the PARIS project. This commit cleans-up the
changes that were added in the previous commit for canary testing.

BUG=b:236394798
TEST=build

Change-Id: I0cf4ce0835f80350b2339328b7e6df168d8fe2da
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3722884
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#49884}