Add common flags to satlab servod start
This PR implements -verbose and -satlab-id flags. It includes a light
refactoring of the servod_container module to facilitate flow needed to
implement verbose flag

This is because the cleanest way to add the functionality to print the
"docker run..." command was to make a function that takes the
ContainerArgs as the input. In the servod_container module we previously
just had one function `StartServodContainer` which took in a set of
string params and transparently created the ContainerArgs and started a
container with them. Now servod_container exposes a method to make
ContainerArgs, and another method to start a container with those
ContainerArgs. This means we have access to the ContainerArgs in order
to print them

Test Plan: Manually ran with verbose flag and observed results. Manually
ran with a satlab ID and observed a custom satlab ID resulted in the
correct behavior

Bug: 232335621
Change-Id: I6d6cf80dad24faa18f8a89e0d70f2d151fa09499
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3735145
Reviewed-by: Prasad Vuppalapu <prasadv@chromium.org>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#50050}