[GoFindit] Improve speed of heuristic analysis for big dependencies
The problem is described in the bug. In this CL, I tried to improve the speed of the dependency analysis by preprocessing the dependencies and putting them in a hashmap with canonical file name as the key. Running the heuristic analysis again only took a few seconds.


Bug: 1341536
Change-Id: Ica14035b7d95c94111896c01a1e6ca7f2872e7df
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3740247
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/main@{#50071}