[Weetbix] Clean-up rules page and fix cypress tests.
Rules are the stateful part of the Weetbix clustering + bug-filing
system that the user can control. Having some way of directly
inspecting that state is useful for debugging and understanding
what the system is doing. This CL migrates the Rules page to
react and links the entries to the rule page.

This CL also fixes cypress tests, which were broken. The only
test that remains broken is the new rule validation test, but
that is because that page is actually broken. It is best fixed
by migrating that page to react.

BUG=1243174
TEST=npm run test

Change-Id: I90c7972d48374219466bbd98cec3aafd3ec89b8a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3747521
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#50080}