[buildbucket] Add a URL return value to the ScheduleLabpackTask method.
As part of my plan to clean up the API of the skylab buildbucket library,
I'm going to remove the method that generates a URL and instead generate
the URL right when the task is created.

We need to do this because the URL depends on the exact name of the task
being created.

BUG=b:230925855
TEST=make test

Change-Id: Iffcc250e54604e49ad018058fa76317aed8dda2a