Remove kweng@google.com from recipes/recipes/infra_repo_trybot.expected/OWNERS.
This suggested change is automatically generated based on group memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot will abandon it. Vote CR +1/2 to approve this change.
See the owner's recent activity for context: https://chromium-review.googlesource.com/q/kweng@google.com

To report an issue, file a bug in the Infra>Security component.

Change-Id: I9f87c99fbeb4eba6004b03c2a235478089f9a0fe