Fix the feature to display and manage default machines
Previously it was difficult to manage default machines in both frontend and backend. To address this, add Default field in the AssetResourceModel to denote whether a given AssetResource is default. Additionally, in the Poros frontend UI, separate default and non-default machines for easier management.

Change-Id: I8ae75d0c8a84afdb082f2aaa4ce818fdb2e03006
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3758351
Commit-Queue: Yunpeng Liu <yunpengl@google.com>
Reviewed-by: Eswaran Krishnamoorthy <eswarank@google.com>
Cr-Commit-Position: refs/heads/main@{#50192}