[recovery] Make ConfigReader private
The ConfigReader API is a little bit awkward at the call site and the only two ways that
we populate a config file are as Base64-encoded strings and as files.

I want to simplify the code at the call site by making this field internal and using methods
to populate it.

TEST=make test

Change-Id: I7219c0484ad24b4bb037deda42c4286279a66281
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3758512
Reviewed-by: Vineet Joshi <vkjoshi@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#50237}