Provide the ability to enable or disable Git-on-Borg retries.
Supporting builders whose configuration is defined in a dependency
project requires accounting for the situation where the builder is
triggered on a revision of the top-level project that has the dependency
project pinned at a revision before the properties file for the builder
was added. In order to do this, it will be necessary to perform
different GoB operations to differentiate between replication lag and a
file truly not existing at the revision and we would not want to wait
until the retries on the initial operation were exhausted.

This changes it so whether or not retries are enabled is determined by
the operation's context. By default, retries are enabled. Calling
gob.DisableRetries will return a new context that when passed to
gob.Execute will not perform retries. Retries can be re-enabled with
gob.EnableRetries. Retries can be disabled for an individual error by
tagging it with gob.DontRetry.

Bug: 1313607
Change-Id: I093c62a4a56cbc4193cd6ae2fbb2bce4d832d9ed