Detect missing properties files for pinned dependencies.
When a builder defined in a dependent project is first added, if the
builder is triggered by a poller watching the top-level project then the
builder will infra fail until the change that added the builder rolls
into the top-level project. This is because the properties file won't
exist at the pinned revision. To prevent that, this change adds
additional logic when downloading the properties file when downloading
from a revision pinned by a top-level project. If the properties file
doesn't exist, the build will be canceled with a summary explaining the
situation.

Bug: 1313607
Change-Id: I3ffd7fc56e8bc15a08ee334ec988615d1e68c1ca