Add Deleted field to ResourceModel and allow users to delete a Resource in the Poros UI
Added Deleted field to ResourceModel to denote whether a Resource should be considered deleted. If a Resource is flagged as deleted, it will not be displayed in the Poros UI.
Implemented a delete button on the Resource creation page that allows the user to delete a given Resource.

Change-Id: I4be5c1f8c1e9cc12d51c4398fc1bba4d23780100
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3761274
Reviewed-by: Eswaran Krishnamoorthy <eswarank@google.com>
Commit-Queue: Yunpeng Liu <yunpengl@google.com>
Cr-Commit-Position: refs/heads/main@{#50264}