Fix inconsistent ending newline in DNS changes
makeNewContent produces an array of strings with no newline at the end
of the lines. It then uses strings.Join(...) to construct a single
string with a newline *in between* each line. This by default results in
no trailing newline for the string then written into a file

However, when adding an entirely new host, that individual line is
created with a newline character, meaning that the string we construct
will have a trailing newline because the last line in the string has a
newline in it already

The end result is that whether the file has a newline or not depends on
whether we are adding a new host or updating it. This has no practical
impact on functionality but was causing issues with unit testing a
future change/can be improved in a simple manner

This removes the newline from any new host, addr line being created, and
always appends a trailing newline to the string we ultimately create and
write to file

BUG=b:200068011
Test Plan=`make test`, `make satlab` and scp to verify `satlab update
dns` still works + produces a consistent file

Change-Id: I9067cdf8a352952be6bbca7323ac89288f066a38
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3763620
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Cr-Commit-Position: refs/heads/main@{#50267}