Add Deleted field in Asset and create delete button in Asset creation page
Added Deleted field to AssetModel and AssetEntity to denote whether an Asset should be considered deleted. If an Asset is flagged as deleted, it will not be displayed in the Poros UI.

Implemented a delete button on the Asset creation page that allows the user to delete a given Asset.

Change-Id: I1370dd934dafd7efbb5176803b06661ce94e0c7e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3763632
Commit-Queue: Yunpeng Liu <yunpengl@google.com>
Reviewed-by: Eswaran Krishnamoorthy <eswarank@google.com>
Cr-Commit-Position: refs/heads/main@{#50292}