upload_debug_symbols: use name from spec instead of filename
This fixes the uploaded name of the debug file to reference the one
inside the .sym file. Previously we were uploading with the .sym
extension by referencing the filename but this is not what the symbol
server expects.

BUG=b:239055291
TEST=go test ./...

Change-Id: I68b1058f44887a5c3b67b2c1467a7795a947edcb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3764661
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Reviewed-by: Nelson Billing <nbilling@google.com>
Auto-Submit: George Engelbrecht <engeg@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>
Cr-Commit-Position: refs/heads/main@{#50323}