[recovery] Create a new version of the no-battery RPM config check action for audit tasks
Based on an in-person discussion, let's do what the python version of labpack does and not
demand a cr50 chip up front in order to perform this test for audit tests.

Deployment tasks will still require a cr50 in order to run the action.

BUG=b:229789619
TEST=make test

Change-Id: Iea72d4e49e8b4a8781f74610154f82393d83c454
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3761701
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50333}