[recovery] Create a new version of the with-battery RPM config check action for audit tasks.
Based on an in-person discussion, let's do what the python version of labpack does and not
demand a cr50 chip up front in order to perform this test for audit tests.

Deployment tasks will still require a cr50 in order to run the action.

BUG=b:229789619
TEST=make test

Change-Id: Ia9a6922320264bc39b66be65c08b863f42d4c51e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3761702
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#50467}