[WIB] Fix source path generation
Generate path based on what the host os is. This simplifies the path
generation logic. Instead of trying to determine in code what type
(linux[/] or windows[\]) path to generate, use inbuilts. This doesn't
work for online windows image builder as we would be technically
running the recipe on a linux host but running jobs on a windows host.
For this we can use a separate path generation logic in execution
module to convert one to other.

Change-Id: I03915b67c8079e81cc402421c211bce4ca9b09c6