Add wheels from https://crbug.com/1344408
Extend the behavior of tpp_libs so that the bin/ directory of each
package is added to PATH in the build environment. This more closely
matches the behavior of 3pp builders, and in this case, allows the
mysqlclient setup script to find mysql_config for build flags. This is
done using env_prefix/env_suffix, which are now implemented for
non-Docker platforms.

Bug: 1344408
Change-Id: I06eac857ff830f32c6bb39416db73351ecce77d6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3763083
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50520}