Correct `satlab get ...` usage printing
`satlab get help dns` incorrectly prints a usage of `satlab dns`
(should be `satlab get dns`). This is because the application we run
the command on has a `Name` of `satlab` and the `get dns` command has a
command of `dns`, When we print the usage, it simply combines the Name
+ command [1], meaning that we just print `satlab dns`

The fix is to change the `Name` of the application `satlab get dns` runs
under. We can do so by reimplementing GetName() to return `satlab get`.
This means usage is now printed as `satlab get` + `dns`. This seems to
roughly follow the approach that the subcommands lib uses [2] and
exactly does what Shivas uses [3]

BUG=b:240151698
Test Plan=`make satlab` and run `satlab get help dns`

[1] https://github.com/maruel/subcommands/blob/main/subcommands.go#L180
[2] https://github.com/maruel/subcommands/blob/7642663bf7467dd2ee5d443a737003516b87f159/sample-complex/ask.go#L14
[3] https://source.chromium.org/chromium/infra/infra/+/main:go/src/infra/cmd/shivas/internal/ufs/cmds/operations/get.go;l=103

Change-Id: I6ff2540e7794851ea16b0b9cf128c0b3e4bab4d6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3787034
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Anh Le <anhdle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50555}