[labpack] Make labpack strict about task names.
As described in the comment in the source code itself, silently falling
back to "recovery" can mask problems.

During the early days of Paris, it made sense to keep going if we encountered
an error inside labpack. Indeed, much of labpack is still written to be as
forgiving as possible.

Since the rollout is nearly complete and labpack and paris have been functioning
well in production for a while now, it no longer makes sense to be this forgiving.
Calling labpack out of contract in new code and having it almost-work in a subtle
and hard-to-detect way is now a larger problem than it was before.

BUG=b:195583069
TEST=make test

Change-Id: I41e93d17f47f482b9355a84bd56e03e4ddb48b4d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3789041
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#50599}