[UFS] Fixing auth issue due to incorrect identity.
CheckFleetTestPolicy RPCs are logging an error "Failed to create identity" due to the addition "user" prefix that is being added when verifying the identity of the test service account that is passed from ctp. CTP test service account is a LUCI identity so no additional prefix is needed.

Bug: b/240371253
Test: make dev; go test;
Change-Id: I492d19b855c3ef190e414eb3c0f0753de4f1677c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3789044
Commit-Queue: Gowri Denduluri <gowriden@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#50606}