[crosadm] Extract information for scheduling units.
Extract information for scheduling units (i.e. multiduts) explicitly.
If for some reason we can't get this information, we fall back to the
GetChromeOSDevice path.

The logic around the GetDeviceData path is as conservative as possible to
avoid messing with the production path for non-multidut devices. We always
make this API call and always log what we received.

BUG=b:240483993
TEST=make test

Change-Id: I1bd5ca00d5b445d393230ba7e269a66dfed1e33b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3792326
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#50662}