[recovery]: Fixed regex to determine the uptime.
The logic to determine whether servo has been up for too long uses a
regex to parse the uptime and extract its value. This regex needed to
be updated to accept the complete range of uptime values.

Also, the parameter to determine the maximum allowed uptime in the
exec has been renamed to better match the intent. Also, this renamed
value now matches the value being set in the config for servo.

Bug=b:240471216
TEST=build, go test ./...

Change-Id: I9123c206f53f1566421808d95ea13642b42dc463
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3798236
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50677}