[Weetbix] Log rule changes.
These changes provide a way of recovering old rules if malicious
or unintended activity occurs. It is anticipated that this situation
is unlikely, so we implement this minimal solution that is easy
to implement but makes recovery harder (as log parsing will be
required to actually do a recovery).

BUG=b:239768873
TEST=Integration tests

Change-Id: Ic784b283fd4824d1601fef7cada83c8a988320eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3807021
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#50705}