[stable_versions] Ignore the firmware section at validation time.
The Python half of the presubmit hook already checks to make sure that the
faft and firmware versions are compatible.

At this point, the firmware version is only used to check the RO version
on the device. It must match the faft version, but is not required to be
built with the OS version.

See b:241439904 comment #8 for details.

Test:

Confirm that './stable_version2 validate-config -remote-file' succeeds.

BUG=b:241439904
TEST=make test

Change-Id: I0039e90b16a349a8ac085805bf08b3f7ba82231c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3814199
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#50784}