[recovery]: Fix the logic about rebooting a long-running servo v3 and labstation.
If servo v3 or the labstation has been running for too long, it can
run into problems.

This commit updates the logic to validate uptime to check for uptime
to be below a threshold, as well as above a threshold. Then this is
used to remedy the cases for too long running servo V3, or too long
running labstations.

Bug=b:240471216
TEST=build, go test ./...

Change-Id: I416704e8123d2fbfddafc194d7aff8a96f67b115
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3814192
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50791}