[crosadm] Scan OS and then browser namespace for multidut devices.
So, the root cause for b:240483993 is that we aren't passing the
namespace to UFS when checking for scheduling units. This CL fixes that.

We check the OS namespace and then the browser namespace.

Test:

I confirmed that this works in an earlier CL (https://chromium-review.googlesource.com/c/infra/infra/+/3821179).

BUG=b:240483993
TEST=make test

Change-Id: I4582645eea4a4a2cb23cecee94057f011390c003
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3821180
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#50828}