recovery: Add ability to recover RO_VPD values
Add functionality to the recovery plans to record VPD values during
deployment and re-flash those values to DUTs during recovery.

The only VPD value included in this CL is "wifi_sar", but it is likely
that others may be needed by other teams, so this implementation makes
adding new VPD keys to track very simple.a

When testing this functionality, I was unable to validate the deployment
step, so instead I temporarily added the "Read RO_VPD data" task to the
recovery step, which showed that the data was properly marshalled into
the new proto map.

BUG=b:240999852
TEST=go clean; go build // In recovery repo
TEST=./mallet local-recovery -dev-jump-host ${labstation} ${host}

Change-Id: Ifcc2a3eff82a85f0f9d24a096162ab65f6b19fd5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3805435
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#50845}