[qemu] Add support to create partition table in disk image
Bug: b/232831070

Use a helper script to create a hard disk image. This is designed to
create a disk image without need for bots to have permission to format
disks. For now we create a MBR disk with a FAT32 partition to copy
necessary dependencies to.

Change-Id: Iae28ee3ccbcf64505c1ba655bdf4828183bc3c9c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3764812
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Adam Todd <actodd@google.com>
Cr-Commit-Position: refs/heads/main@{#50849}