[WIB] Refactor code - use module object, refactor comments
Pass the module object to all the objects that use any module. This
replaces the previous approach which passed induvidual modules,
whenever/wherever required. This should simplify reading the code and
implementing new code as we don't have to modify multiple source files
to use a new module.

Also modifies the comments on the functions to make it render properly
in README.recipes.md

Change-Id: Id7d9969eccd8f45b2d4c9b8b00a00ee3a3bf1c9b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3788688
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50884}