[WIB] Add support to safely shutdown vm
Bug: b/182061277

Shutdown VM safely with best attempt. Guest Windows OS running on VM
needs to shutdown after the customization is applied. This is done by
running `Stop-Computer` powershell expression. However, this can also
be done using the configs. The following situations are possible once
the vm has finished applying the configs

1. Config ran a `Stop-Computer` VM is shutting down
2. Config didn't run `Stop-Computer` VM is running
3. Powershell client PSOverCom is dead, but VM is running
4. Last action lead to VM reboot and VM is running

The algorithm to safely shutdown the vm is as follows.

1. Attempt to send `Stop-Computer` signal. This will fail if powershell
   client PSOverCom is dead or VM is dead or rebooting.
2. Attempt to send powerdown signal(Power button) to the VM. Sometimes
   (WinPE) OS ignores this signal.
3. Attempt to Force kill the VM by asking QEMU to kill the VM

If we kill the VM, the recipe execution fails as we cannot be sure
that the disks are properly written to and the OS exited safely.

Change-Id: I500b2378aa667af90624d87c2dc0dc630e31c9c3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3814197
Reviewed-by: Adam Todd <actodd@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50895}