recovery: Fix a timing issue that may cause Android repair failure
After switch adb to root mode, the device may be in a flaky state
for a short period and it has chance to cause actions right after
it fails due to no device found. This CL add a short sleep after
switch adb to root mode to address this flakiness.

BUG=b:241604162
TEST=go test ./... and repeat 10 successes with mallet local-recovery

Change-Id: Ib2db2ba33c676889b36978e121a95f176ed79705
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3828481
Commit-Queue: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Sergey Fetisov <sfetisov@google.com>
Cr-Commit-Position: refs/heads/main@{#50901}