recovery: Fork a download usb action that less strict
This is bug that may cause servo usbkey is not detectable
from labstation side but can still be seen on DUT side.
So in the repair we want to be less strict and try boot
DUT into servo usbkey no matter if labstation side check
is success or not. However, during deployment we want to
ensure that usbkey is in good state and an expected image
is loaded.

BUG=b:201796172
TEST=go test ./... and run mallet

Change-Id: Iebfa7324e0197e1d0cc880ded27da85416d597d3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3832920
Reviewed-by: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Garry Wang <xianuowang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#50964}