ufs: modify HWID related protos and methods for clarity
Please refer to this doc for context: go/2022q3-hwid-proto-refactor. The
changes to the proto enhances usability and maintainability of the
HwidData proto. The methods associated are also changed as well. UFS
datastore will gradually be storing the new HwidData messages as well.

When parsing the HwidDataEntity, it will first try to get the HwidData
proto. If it fails to find any DutLabels, it will fall back to try
parsing the entity as a DutLabel proto. This will be removed in time
once the entities are overwritten with HwidData in the data blob.

BUG=b/239589687
TEST=go test ./...; make dev

Change-Id: I784cf45d0b6df6625c5d57f909a305c164db25d7