[WIB] Fix git src if we are using it in debug mode
Bug: None

This should have accompanied the debug mode CL. However, this is a quick
fix to unblock config development. Sometimes the refs in git can be of
format `refs/changes/123456/7`. When this is downloaded, as windows uses
`\` as path separator this is changed. Ensure that the right path is
generated by the git src module.

Change-Id: I09147a4940e387a39306bcff044b024b77974ad1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3843150
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Adam Todd <actodd@google.com>
Reviewed-by: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#51063}