[croasdm] Drop the ChromeOSDeviceData RPC call from the UFS interface
As discussed earlier today, this RPC call is one of the few RPC calls
that ignores the reported namespace because it only applies to the OS
labs.

This type of RPC should be removed from UFS probably, so let's prevent
crosadm from depending on it.

BUG=b:240483993
TEST=make test

Change-Id: I5f1e68be7e93422bbaa6cdf3c51aaa1f5394bfbc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3855403
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#51116}