Fix gn targets in gn file arguments.
Gn targets may appear in BUILD.gn as arguments in form of //some_target
or //some:target, and may be considered as a path that should be fixed.

Add a special handling for such arguments and do not consider them as an
error in potential path but skip fix attempt and return them as they
are.

Change-Id: I53652122efdf509bedb98d6f7c60d993ea96a606
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3865308
Reviewed-by: Gavin Mak <gavinmak@google.com>
Commit-Queue: Artem Sumaneev <asumaneev@google.com>
Cr-Commit-Position: refs/heads/main@{#51233}