[karte] Remove datastore wrapper library.
Trying to "insulate" Karte from the LUCI datastore library (which is itself a wrapper
around BigQuery's library) is pointless and leads to more code complexity for insufficient
benefit.

Also, datastore already has an in-memory implementation that can be used in unit tests
easily, so the datastore wrapper is not giving us the ability to write unit tests with better
coverage than we would otherwise have.

BUG=None
TEST=make test

Change-Id: Ie6f7ab465ab34d48821c4dbe828a33a3560ef1ad
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3872251
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#51318}