Implement CBI is Present Action
Implements an action to check if the DUT being repaired contains CBI contents. See go/cbi-auto-recovery-dd for more background. This is still a NO-OP as there is not critical action connecting this action to any plan.

This change was tested by running the unit tests for Paris and by running a custom plan https://paste.googleplex.com/6200104330461184 that only contains the CBI actions against a DUT that contains CBI and a dut that DOES NOT contain CBI to see if it correctly marked the DUT as having CBI present or not. See go/working-with-paris for what commands were used to build and run these changes.

BUG=b/235000813
TEST=manual and unit

Change-Id: I10109e30e8dfb4a7962db211825a658e1075847b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3862893
Commit-Queue: Sean Carpenter <seancarpenter@google.com>
Reviewed-by: Sean Carpenter <seancarpenter@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Ashley Wans <acwans@google.com>
Cr-Commit-Position: refs/heads/main@{#51372}