[crosadm] Change the dev environment to use GAEv2
The CrOSSkylabAdmin project has been using go116-compatibility for a long time.
It's time to remove this temporary solution. In order to facilitate the transition
to being an ordinary GAEv2 application, move the dev environment (and only the dev
environment) over to GAEv2 mode.

BUG=b:173543443
TEST=make up-staging && make switch-staging

Change-Id: I82f57dfa93724521b3486161535e0219bc7b3908
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3892975
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#51500}