[ufs] Make client close errors fatal in pubsub_component.
staticcheck has correctly determined that we are ignoring the error output
of client.Close().

So, let's add a defer block that handles the status.

For now, I am making the error fatal and making it cause the process to terminate immediately
so that an alert for a failed dumper task is triggered higher in the stack.

BUG=b:230200662
TEST=make test

Change-Id: I28b7298cb52b4331bf3a5b522cd69aa5937e592f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3897499
Reviewed-by: Julio Hurtado <juahurta@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#51587}