myjobs: Get current builder props from led
We will need to override certain module properties without overriding
other props of the same module: for example, we should be able to set
$chromeos/orch_menu.filter_child_builds without changing the rest of
$chromeos/orch_menu's properties as defined in infra/config. Due to
https://crbug.com/1347511, this is not currently possible via `bb`.

As a workaround, we can fetch the builder's properties as defined in
infra/config via `led get-builder`, override them on a per-property
basis, and then call `bb add` with the full property JSON.

This CL adds authentication for the `led` CLI, adds a function to fetch
and parse input properties via `led get-builder`, and writes them to a
tempfile. Also adds unit tests. The authentication work involves a
little refactoring.

BUG=b:240289098
TEST=go test .
TEST=Comment out the line to delete the tempfile, run the CLI, and read
the tempfile to ensure the input properties look OK

Change-Id: I0cd7bb9cb244ecbb7fbbe0b8a407062701dbe37e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3908162
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Chris Gerber <gerb@google.com>
Cr-Commit-Position: refs/heads/main@{#51708}