[ethernethook] [mallet] Add LsSmall
Add the ability to synchronously get a small number of object descriptors.
This method, LsSmall, will later be used to get a list of all paths associated with an
object as the first step towards grabbing the control file, the status.log file, etc.

BUG=b:244478738
TEST=make AND ./mallet ethernet-hook -prefix /path/to/test

Change-Id: I46240af785b647d1bd717ec223e7ee623ba0a5d8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3908154
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#51717}