Don't set the gitiles commit in the input unless it was already set.
Currently, the bootstrapper will set the gitiles commit on the build to
the commit for the dependency repo if the config is present in a
dependency repo. In the case of such a builder that doesn't actually
check out the dependency repo (e.g. chrome/isolation/win-chrome-secops
which is defined in src-internal but does not check it out), this will
cause bot_update to fail because the gclient config won't contain a
repo_path_map entry for the dependency repo. This changes it so that an
existing input commit will be modified with the retrieved revision. The
commit information is still transmitted to be used by
chromium_bootstrap.update_gclient_config which has been updated to
account for this situation.

Change-Id: I05a0d07f4ff386ae5cc3e0e7a0c10d66a2efeb7b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3914194
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#51746}